Start on GetMapping request to view MovieReview

+++ media-wrangler-back-end/src/main/java/com/mediawrangler/media_wrangler/data/MovieReviewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovieReviewRepository extends JpaRepository<MovieReview, Integer> {
+public interface MovieReviewRepository extends JpaRepository<MovieReview, Long> {
 
 }
